+++ pkg/model/model.go
@@ -33,6 +33,11 @@ type MaskContextEngine interface {
 	MaskContext(Dictionary, string, ...Dictionary) (Dictionary, error)
 }
 
+// HasCleaner interface provides a function to apply on cleanup
+type HasCleaner interface {
+	GetCleaner() FunctionMaskContextEngine
+}
+
 // FunctionMaskEngine implements MaskEngine with a simple function
 type FunctionMaskEngine struct {
 	Function func(Entry, ...Dictionary) (Entry, error)
@@ -41,6 +46,16 @@ type FunctionMaskEngine struct {
 // Mask delegate mask algorithm to the function
 func (fme FunctionMaskEngine) Mask(e Entry, context ...Dictionary) (Entry, error) {
 	return fme.Function(e, context...)
+}
+
+// FunctionMaskContextEngine implements MaskContextEngine with a simple function
+type FunctionMaskContextEngine struct {
+	Function func(Dictionary, string, ...Dictionary) (Dictionary, error)
+}
+
+// MaskContext delegate mask algorithm to the function
+func (fme FunctionMaskContextEngine) MaskContext(e Dictionary, key string, context ...Dictionary) (Dictionary, error) {
+	return fme.Function(e, key, context...)
 }
 
 type MaskFactory func(Masking, int64, map[string]Cache) (MaskEngine, bool, error)
@@ -109,6 +124,7 @@ type TemplateEachType struct {
 
 type MaskType struct {
 	Add               Entry                `yaml:"add,omitempty" jsonschema:"oneof_required=Add"`
+	AddTransient      Entry                `yaml:"add-transient,omitempty" jsonschema:"oneof_required=AddTransient"`
 	Constant          Entry                `yaml:"constant,omitempty" jsonschema:"oneof_required=Constant"`
 	RandomChoice      []Entry              `yaml:"randomChoice,omitempty" jsonschema:"oneof_required=RandomChoice"`
 	RandomChoiceInURI string               `yaml:"randomChoiceInUri,omitempty" jsonschema:"oneof_required=RandomChoiceInURI"`
@@ -138,7 +154,8 @@ type MaskType struct {
 
 type Masking struct {
 	Selector SelectorType `yaml:"selector"`
-	Mask     MaskType     `yaml:"mask"`
+	Mask     MaskType     `yaml:"mask" jsonschema:"oneof_required=Mask"`
+	Masks    []MaskType   `yaml:"masks" jsonschema:"oneof_required=Masks"`
 	Cache    string       `yaml:"cache,omitempty"`
 }