+++ pkg/model/pipeline.go
@@ -20,7 +20,6 @@ package model
 import (
 	"errors"
 	"io/ioutil"
-	"strconv"
 	"time"
 
 	"github.com/goccy/go-yaml"
@@ -65,54 +64,72 @@ func BuildCaches(caches map[string]CacheDefinition, existing map[string]Cache) m
 
 func BuildPipeline(pipeline Pipeline, conf Definition, caches map[string]Cache) (Pipeline, map[string]Cache, error) {
 	caches = BuildCaches(conf.Caches, caches)
-	for _, v := range conf.Masking {
+	cleaners := []Processor{}
+	for _, masking := range conf.Masking {
 		nbArg := 0
 
-		if v.Mask.FromCache != "" {
-			cache, ok := caches[v.Mask.FromCache]
-			if !ok {
-				return nil, nil, errors.New("Cache '" + v.Cache + "' not found for '" + v.Selector.Jsonpath + "'")
-			}
-			pipeline = pipeline.Process(NewFromCacheProcess(NewPathSelector(v.Selector.Jsonpath), cache))
-			nbArg++
-		}
+		allMasksDefinition := append([]MaskType{masking.Mask}, masking.Masks...)
 
-		for _, factory := range maskFactories {
-			mask, present, err := factory(v, conf.Seed, caches)
-			if err != nil {
-				return nil, nil, errors.New(err.Error() + " for " + v.Selector.Jsonpath)
+		for _, maskDefinition := range allMasksDefinition {
+			virtualMask := Masking{
+				Selector: masking.Selector,
+				Mask:     maskDefinition,
+				Masks:    nil,
+				Cache:    masking.Cache,
 			}
-			if present {
-				if v.Cache != "" {
-					cache, ok := caches[v.Cache]
-					if !ok {
-						return nil, nil, errors.New("Cache '" + v.Cache + "' not found for '" + v.Selector.Jsonpath + "'")
+
+			if virtualMask.Mask.FromCache != "" {
+				cache, ok := caches[virtualMask.Mask.FromCache]
+				if !ok {
+					return nil, nil, errors.New("Cache '" + virtualMask.Cache + "' not found for '" + virtualMask.Selector.Jsonpath + "'")
+				}
+				pipeline = pipeline.Process(NewFromCacheProcess(NewPathSelector(virtualMask.Selector.Jsonpath), cache))
+				nbArg++
+			}
+
+			for _, factory := range maskFactories {
+				mask, present, err := factory(virtualMask, conf.Seed, caches)
+				if err != nil {
+					return nil, nil, errors.New(err.Error() + " for " + virtualMask.Selector.Jsonpath)
+				}
+				if present {
+					if virtualMask.Cache != "" {
+						cache, ok := caches[virtualMask.Cache]
+						if !ok {
+							return nil, nil, errors.New("Cache '" + virtualMask.Cache + "' not found for '" + virtualMask.Selector.Jsonpath + "'")
+						}
+						switch typedCache := cache.(type) {
+						case UniqueCache:
+							mask = NewUniqueMaskCacheEngine(typedCache, mask)
+						default:
+							mask = NewMaskCacheEngine(typedCache, mask)
+						}
 					}
-					switch typedCache := cache.(type) {
-					case UniqueCache:
-						mask = NewUniqueMaskCacheEngine(typedCache, mask)
-					default:
-						mask = NewMaskCacheEngine(typedCache, mask)
+					pipeline = pipeline.Process(NewMaskEngineProcess(NewPathSelector(virtualMask.Selector.Jsonpath), mask))
+					nbArg++
+				}
+			}
+
+			for _, factory := range maskContextFactories {
+				mask, present, err := factory(virtualMask, conf.Seed, caches)
+				if err != nil {
+					return nil, nil, errors.New(err.Error() + " for " + virtualMask.Selector.Jsonpath)
+				}
+				if present {
+					pipeline = pipeline.Process(NewMaskContextEngineProcess(NewPathSelector(virtualMask.Selector.Jsonpath), mask))
+					nbArg++
+					if i, hasCleaner := mask.(HasCleaner); hasCleaner {
+						cleaners = append(cleaners, NewMaskContextEngineProcess(NewPathSelector(virtualMask.Selector.Jsonpath), i.GetCleaner()))
 					}
 				}
-				pipeline = pipeline.Process(NewMaskEngineProcess(NewPathSelector(v.Selector.Jsonpath), mask))
-				nbArg++
 			}
 		}
-
-		for _, factory := range maskContextFactories {
-			mask, present, err := factory(v, conf.Seed, caches)
-			if err != nil {
-				return nil, nil, errors.New(err.Error() + " for " + v.Selector.Jsonpath)
-			}
-			if present {
-				pipeline = pipeline.Process(NewMaskContextEngineProcess(NewPathSelector(v.Selector.Jsonpath), mask))
-				nbArg++
-			}
+		if nbArg == 0 {
+			return pipeline, nil, errors.New("No masks defined for " + masking.Selector.Jsonpath)
 		}
-		if nbArg != 1 {
-			return pipeline, nil, errors.New("Not the right number of argument for " + v.Selector.Jsonpath + ". There should be 1 and there is " + strconv.Itoa(nbArg))
-		}
+	}
+	for _, cleaner := range cleaners {
+		pipeline = pipeline.Process(cleaner)
 	}
 	return pipeline, caches, nil
 }