+++ pkg/templatemask/templatemask.go
@@ -19,39 +19,20 @@ package templatemask
 
 import (
 	"bytes"
-	"strings"
-	"text/template"
-	"unicode"
 
-	sprig "github.com/Masterminds/sprig/v3"
 	"github.com/cgi-fr/pimo/pkg/model"
+	"github.com/cgi-fr/pimo/pkg/template"
 	"github.com/rs/zerolog/log"
-	"golang.org/x/text/runes"
-	"golang.org/x/text/transform"
-	"golang.org/x/text/unicode/norm"
 )
 
 // MaskEngine is to mask a value thanks to a template
 type MaskEngine struct {
-	template *template.Template
-}
-
-// rmAcc removes accents from string
-// Function derived from: http://blog.golang.org/normalization
-func rmAcc(s string) string {
-	t := transform.Chain(norm.NFD, runes.Remove(runes.In(unicode.Mn)), norm.NFC)
-	result, _, _ := transform.String(t, s)
-	return result
+	template *template.Engine
 }
 
 // NewMask create a MaskEngine
 func NewMask(text string) (MaskEngine, error) {
-	funcMap := template.FuncMap{
-		"ToUpper":  strings.ToUpper,
-		"ToLower":  strings.ToLower,
-		"NoAccent": rmAcc,
-	}
-	temp, err := template.New("template").Funcs(sprig.TxtFuncMap()).Funcs(funcMap).Parse(text)
+	temp, err := template.NewEngine(text)
 	return MaskEngine{temp}, err
 }